pyproject.toml
[tool.poetry]
name = "liis_test"
version = "0.1.0"
description = ""
authors = ["German Gritsenko <gerfest776@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
pre-commit = "^2.18.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
gunicorn = "^20.1.0"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.1.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
